feat: add DataBaseException and update ResouceExceptionHandler and UserService

+++ src/main/java/com/exa/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.exa.entities.User;
 import com.exa.repositories.UserRepository;
+import com.exa.services.exceptions.DataBaseException;
 import com.exa.services.exceptions.ResourceNotFoundException;
 
 @Service // @Component serviria para registrar a classe como componente do Spring, porem
@@ -35,7 +38,26 @@ public class UserService {
 
     // metodo pra deletar um usuario
     public void delete(Long id) {
-        repository.deleteById(id);
+        try {
+            // precisa agora manualmente fazer isso pois o JPA nao lança mais exception caso o id nao exista
+            if (!repository.existsById(id)) {
+                throw new EmptyResultDataAccessException(1);
+            }
+            repository.deleteById(id);
+        // aqui vai o tipo mais genérico de exceção, pra pegar qualquer exceção que o deleteById possa lançar e printar o erro especifico que voce quer pegar pra depois colocar aqui no lugar do RuntimeException
+        // } catch (RuntimeException e) {
+        // a maneira usada aqui pra ir tratando excecoes é imaginar cenarios de erro e ir tratando um por um, como este primeiro que é caso alguem tente deletar um usuario que nao existe
+        } catch (EmptyResultDataAccessException e) {
+            // printStackTrace apenas para mostrar o erro no console
+            throw new ResourceNotFoundException(id);
+        // este seria o proximo cenario de erro, caso o usuario tenha alguma dependencia que nao pode ser deletada, testando primeiro sempre com:
+        // } catch (RuntimeException e) {
+        //     e.printStackTrace();
+        // }
+        // e depois de pegar o erro especifico, colocar aqui no lugar do RuntimeException e lançar o erro próprio criado
+        } catch (DataIntegrityViolationException e) {
+            throw new DataBaseException(e.getMessage());
+        }
     }
 
     // metodo pra atualizar um usuario
